refactor tests: split add command tests
